convert request to prompt
