--- app/Http/Requests/TranslationRequest.php
@@ -1,37 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace App\Http\Requests;
-
-use App\Enums\Tone;
-use Illuminate\Foundation\Http\FormRequest;
-use Illuminate\Validation\Rules\Enum;
-
-class TranslationRequest extends FormRequest
-{
-    /**
-     * Determine if the user is authorized to make this request.
-     */
-    public function authorize(): bool
-    {
-        return true;
-    }
-
-    /**
-     * Get the validation rules that apply to the request.
-     *
-     * @return array<string, \Illuminate\Contracts\Validation\Rule|array|string>
-     */
-    public function rules(): array
-    {
-        return [
-            'word' => ['required', 'string'],
-            'context' => ['string'],
-            'input_language' => ['required', 'string'],
-            'output_language' => ['required', 'string'],
-            'tone' => ['required', new Enum(Tone::class)],
-            'audience' => ['string'],
-        ];
-    }
-}
